Add guarded startup path for IT staff backfill
