+++ instrumentation.ts
@@ -1,0 +1,7 @@
+export async function register() {
+	if (process.env.NEXT_RUNTIME !== "nodejs") {
+		return;
+	}
+
+	await import("./instrumentation-node");
+}
