+++ package.json
@@ -15,6 +15,7 @@
 		"db:seed:showcase": "bun prisma/seed.dev-showcase.ts",
 		"db:e2e:seed": "bun prisma/seed.e2e.ts",
 		"db:backfill:it-staff": "bun scripts/backfill-it-staff-users.ts --apply",
+		"db:backfill:it-staff:dry-run": "bun scripts/backfill-it-staff-users.ts",
 		"db:studio": "bunx prisma studio",
 		"db:reconcile-checksums": "bun scripts/reconcile-migration-checksums.ts",
 		"test": "bun run test:unit",
